Set sqitch deploy mode to "change"
This lets successful changes earlier in a deploy remain unreverted when
a subsequent change fails.  Particularly useful when changes may be
missing revert scripts, which I expect to be common with data-only
changes since upserts make it hard to define sane revert behaviour (and
I expect upserts to be common).