Add reportable conditions view
Create a view in the `shipping` schema that comprises all samples with
positive test results for reportable conditions (e.g. measles).

In this view, track also the status of whether or not the sample has
been reported. Curently the plan is to report samples via Slack to the
pertinent channels.