wip! tract to puma
Data files and code for mapping tracts to PUMAs, but not yet integrated
into the warehouse.location.hierarchy of tracts.

The sticking point was that `id3c location import` doesn't expose a way
to make hierarchy dynamic per feature without modifying the GeoJSON (and
I wanted to avoid modifying the GeoJSON so that we ingest vanilla Census
data).

There are several possible solutions, but I think the most expedient one
(and one that doesn't modify the tract GeoJSON) is to execute a SQL
script after tract impot which reads the tract → PUMA mapping file and
updates tract hierarchies.

A more general solution would be to extend `id3c location import` to
support a feature like this, but I'm not sure what the CLI options for
that best look like.