etl redcap-det: use full URL references in FHIR... Resources
If we want to use `fhir.resource` resolution methods, we need to use a
literal or full URL reference instead of a logical reference (e.g.
identifier).
Update the declaration of each FHIR Resource so that a `fullUrl` key is
accessible by other Resource-generating methods.

We may want to consider dropping identifiers altogether from reference
objects if we keep this new approach.