Change immunization to questionnaire response item
Immunization resource references the Encounter where the immunization
happened, but does not have a reference for when the immunization was
reported. We are removing the Immunization resource and replacing
it with an item in the QuestionnaireReponse. The answer must have
"valueBoolean" but can also have "valueDate" if the date of the
immunization is known.