clinical parse-uw: Use collection date preferentially to lab date
Collection date is the more accurate date for our purposes, but in
previous datasets it wasn't always available.  Lab date was always
available, but, as I realized now, is sometimes is a day or even week
later.  A day of fudge is ok for our work, but a week probably isn't.