etl clinical: Fix conversion of scalars into single-element lists
This was added as part of the recent SCH updates, but wasn't triggered
by any of the SCH records.  In working on UW updates, the conversion
failed to take place because of the faulty conditional, and an exception
for an unknown insurance type was then thrown a few lines later as the
code tried to iterate over a string.