etl redcap-det: Rework identifier hashing
Refactors logic into both our shared ….command.etl.fhir module and the
core ….command.de_identify.generate_hash() function for consistent and
robust behaviour.

The order of personal information in the hash is consistent and no
longer reliant on stringification of dict objects.  All information is
required and must not be blank so as not to compromise the integrity of
the hash.

The hashing secret is provided by the now-standard
ID3C_DEIDENTIFY_SECRET environment variable, which is the core
function's default.  PARTICIPANT_DEIDENTIFIER_SECRET is a holdover from
our custom clinical command, and there it is preserved for backwards
compat with previous datasets.