Add view shipping.fhir_encounter_details_v1
This is a custom view that works specifically for `encounter.details`
that contain FHIR resources generated for Seattle Flu. It pulls out
the details into a view that can be easily joined with other views
using `encounter_id`.

This view only pulls out "core" questions from `QuestionnaireResponses`.
Other questions can be added to the view in the future.