redcap-det: Fallback to record id-based individual identifier generation
Instead of special casing only zip code and bailing out if other pieces
of information are missing, take a best-effort approach and assume the
record represents a unique participant who hasn't participated
previously and won't participate again in the future.

This is a broadening of the logic we previously used for zip code, and I
think it makes more sense in the bigger picture.  I only realized it
when I was digging into noisy logs, though!  Those same noisy logs are
silenced since missing personal info is now less of an issue with this
more general fallback.