etl redcap-det kiosk: Distinguish between UW and SCH SeaMar
If the staff group of the employee using the kiosk app works at Seattle
Childrens Hospital, then the kiosk site is Seattle Childrens SeaMar.
Otherwise, it is UW SeaMar. Previously we were not distinguishing
between these two sites, and we were assigning all "SeaMar" kiosks to
the `UWSeaMar` site in warehouse.