views: Add best available encounter date & season
Add a `best_available_encounter_date` column to the new view that
captures the best available site information for all samples. This is
similar to what we did in `shipping.metadata_for_augur_build_v3`.

Also, add a `season` column that shows "Y1" or "Y2" depending on the
best available encounter date.

Now that we're not only handling site-specific data, rename the
auxiliary query aliases and the view.