redcap-det swab-and-home-flu: Add home-flu specific changes
The most major change is changing the encounter site based on the
record's `study_type`.

The two studies share a lot of the same field names so a majority of
the ETL process did not need to be changed. In cases where the same
data is provided under different field names, I just used the `or`
operator in the variable assignment.