Remove anachronistic grant
A past rework of roles/augur-build-exporter/grants in "roles: Update
augur-build-exporter for v3" (5ea943f) accidentally modified both the
…/grants.sql script and the …/grants@2020-01-31.sql script.  It should
have modified just the former.

This anachronism didn't cause issues at the time because the SQL script
here was never run, but it errors when deploying from scratch or from a
much older point.

It's an easy mistake to make, and a better CI setup would have caught
this for us before merge by

    a) ensuring deploy from scratch works, and/or
    b) ensuring reworks copy X.sql → X@TAG.sql without any changes.