Fix revert and verify scripts for latest shipping/views rework
These worked just fine on my local development environments, but failed
with permission denied errors on id3c-production because the "postgres"
user we deploy with in RDS isn't a true superuser (there are no true
superusers in RDS).

Since the owner of shipping.flu_assembly_jobs_v1 is now reporter, not
postgres, the revert script must first alter the owner back to postgres
(the current user) before it can issue create or replace view.  Altering
a view's owner requires the user is a member of the new owning role and
has CREATE on the view's schema.  The postgres user does own the
shipping schema.

Similarly for the verify script, because postgres isn't actually a
superuser, once it is no longer the owner of that view it can't
automatically select from it.  The postgres can, however, switch to the
reporter role.

I verified these changes now work by issuing `sqitch rebase testing
@2020-03-05b`.