Create a "view-owner" role
Making "reporter" the owner of some views was a mistake on my part.  I
failed to remember that doing so would allow any member of the role,
i.e. login users granted "reporter", to replace, drop, or otherwise
manage the views!  Instead, make a new "view-owner" role which is also a
member of "reporter", but which will not have any members itself.

Because the changes since sqitch tag @2020-03-05b are not actually
deployed (well they were, but I reverted them), retroactively modify the
previous changes and insert the "view-owner" appropriately.  This avoids
some ugly reworks which would be only for the ceremony of it.