etl redcap-det scan: Add SCAN project changes
The main change is now we are using the raw coded values for multiple
choice answers instead of the label string. This is to allow easier
mapping for translated versions of the same survey.

Currently ingests the raw coded value for QuesionnaireResponse
items, but this can be updated to decode the values.