etl: Dynamically accept FHIR encounter status
Instead of always hard-coding the required FHIR Encounter.status to
'finished', accept an argument for encounter status with a default of
'finished'.

In the future, especially if we start to rely more on retrospetive data,
we may want to consider removing this default and updating our ETLs to
always provide an Encounter status.