Rework joins in shipping.hcov19_observation_v1 to prevent unintended output rows
The two joins across one-to-many relationships are pushed into
subqueries that handle collapsing to one-to-one relationships.

Encounter-related locations now use the new "primary" encounter-location
concept introduced in core ID3C to pick the "best" available location.
This will usually be a residence, but may be lodging or a workplace if
no residence is available.

hCoV-19 presence/absence results are explicitly collapsed in a CTE prior
to joining with an explicit preference for positive results over
negative ones.  This could be extended in the future to also include a
device/source/method preference.

Additionally, the previous target → organism left join is made an inner
join in the subquery, and that subquery *as a whole* is then left
joined.  This resolves the issue of the previous "lineage is null"
condition including p/a results which don't refer to an organism when
the intent was just to include samples without hCoV-19 results.

Similarly, the condition restricting samples to "tested or after date X"
now uses the subquery's composite (row) type ("hcov19_presence_absence
is distinct from null") instead of any individual column, which might be
validly null even if the join produced a matching row.

Finally, the exclusion of environmental samples is made safe against the
(unlikely but possible) case of sample.details being null.

Other minor changes are also included, such as tweaking output column
names to more closely match those we've used elsewhere.