etl: Note usage of Encounter.period violates FHIR
Our current usage of the Encounter.period start and end dates violate
the FHIR docs, because these dates are supposed to cover the period of
a hospitalization (if applicable).
There is too much downstream (ETL FHIR) consideration for this to be
resolved at this time, so mark this as a `FIXME` and put it aside.