etl redcap-det scan-en: remove `illness_questionnaire` from required
From Misja/Izzy: “Sometimes people don't hit "complete" on the survey,
and it's marked as "incomplete".

We want to ingest a record even if this happens!

The illness_questionnaire only contains responses for the
QuestionnaireResponse resource, which is not required to create an
Encounter. Since we already have code in place to prevent empty
responses for the QuestionnaireResponse, I don't see any harm in
removing the `illness_questionnaire` from the list of required
instruments.