etl redcap-det scan: Ingest 'able_to_test'
If a sample arrives in a condition that is unable to be tested, it will
be marked never_tested at unboxing.

If later during lab testing/processing it falls into the never_tested
category (such as a PCR fails), the REDCap record will be updated w/
"no" for the able_to_test variable.

If a sample is unable to be tested, add an anotation to the
Specimen.note with text "never-tested".