etl redcap-det scan: add follow-up Questionnaire Response
Ingests all fields from the follow-up questionnaire that are not
free-text fields.

Included questions regarding symptoms in the Questionnaire Response
because the questions are comparing how the symptoms have changed over
the last 7 days, not just if they have the symptoms.

Also note that this follow-up questionnaire includes questions about
other individuals in the participant's household. Since these household
members are not officially enrolled in the study and we do not an
official encounter with them, we are just including these questions
in the Questionnaire Response for the original participant.