shipping.scan_encounters_v1: filter out follow-up encounters
This view should only contain the initial encounter and all of
the questionnaire responses from the initial questionnaire.
A separate view will be created specifically for the follow-up
encounters and corresponding questionnaire responses.

Creating two separate views that are specific to type of encounter
makes linking the initial encounter and follow-up encounter less
complicated.