shipping.fhir_questionnaire_responses_v1: add unique index
Add unique index on `encounter_id`&`link_id` combination. A unique
index is required for a materialized view to be able to be refreshed
with the `concurrently` parameter. This ensures that users are not
locked out of the view while it's refreshing. This allows us to refresh
the view at a more regular interval without interfering with downstream
user access.