shipping views: drop all views at the top in order of dependency
Drop all views at the top in order of dependency so that we don't have
to worry about inter-view dependencies when reworking view definitions.

This did require re-ordering of view definitions so that they can be
created in the order of dependencies as well. Also took this opportunity
to organize the views into categories instead of having them scattered
across the file.