etl redcap-det scan: skip "false" enrollments
The study only wants us to ingest "true" enrollments where a participant
was mailed a swab kit. "True" enrollments can be identified by the
participant completing the illness_questionnaire. However, there is
a bug in REDCap that sometimes leaves the instrument marked as
incomplete/unverified. To get around this, we are checking that the
`illness_q_date` is filled in to verify that the instrument is complete.
However, there is yet another bug in REDCap that sometimes leaves the
`illness_q_date` blank. To catch these records, we are also checking
that the back_end_mail_scans instrument is complete.

This leaves us with a majority of records being ingested the day of
enrollment due to the completion of the `illness_q_date` and some
straggling records ingested the day after enrollment where the
back_end_mail_scans instrument is complete.