redcap-det scan: Assign site according to `location_type` if available
The new SCAN In-Person Enrollments REDCap project has a `location_type`
field that indicates the site of the kiosk. Assign site according to
this field if available. All other SCAN encounters are still using the
`SCAN` site.

Currently only has `UWGreek` as a new site. The REDCap project does have
a `daycare` option for the `location_type` field, but that is not being
used yet and we have no idea what site that will represent.