etl redcap-det scan: update handling of specimen barcode
Always uses `return_utm_barcode` if that field exists in the REDCap
record.

In-person enrollments don't have this field, so use the
`utm_tube_barcode_2` field. If this field is blank due to scanning error,
then use the manually entered barcode. If the manually entered values
for `reenter_barcode` and `reenter_barcode_2` don't match then return
None to err on the side of caution.