shipping.scan_return_results_v1: Use sample.collected going forward
…instead of the encounter date.  The collect_ts column (as the name
implies) is labeled as the collection date in the result PDFs, and
that's what's more useful to report.

Imposes a date cutoff based on the encountered date to avoid
retroactively changing results for previous samples.  We used
encountered date because it's an easy date to communicate to the other
study teams as the switchover date.