age_ceiling: update type hint for `age` parameter and return value
The `age` parameter can be either a float or an int and the return type
would be determined by the `age` parameter type.

It was initially just set to float because in PEP 484 type hints says:
> this PEP proposes a straightforward shortcut that is almost as
effective: when an argument is annotated as having type
float, an argument of type int is acceptable

However, this resulted in confusing mypy errors in the use of
`age_ceiling` for the SCAN REDCap ETL. Update both to prevent future
confusions.