Add `first_record_instance` decorator for REDCap DET ETLs
This is a decorator that removes the `record_instances` argument from
routine's kwargs and only passes the first REDCap record from the
`record_instances` as the `redcap_record` kwarg to the routine.

Logs a warning if there are multiple instances of a record in the list
of `record_instances`.

Created for an easy way to make the change of passing `record_instances`
instead of `redcap_record` to custom REDCap DET ETLs backwards
compatible with existing ETLs.

Use the new decorator on all existing REDCap DET ETLs that are only
expecting one record instance per record id.