etl redcap-det scan: Update how to determine `specimen_received`
The SCAN Husky project only uses post_collection_data_entry_qc to mark
"never-tested". So we need to rely on `nasal_swab_collection` instrument
to know that we have sample data to ingest.
Only rely on `nasal_swab_collection` if the record is from a kiosk
project, where the record does not have the `back_end_mail_scans`
instrument.