uw-reopening ETL: Check that encounter data exists
Adds a check to the start of the record instance loop to make sure that
we have the data we need to create an ENCOUNTER encounter.
Otherwise, skip the record instance with a log entry at the DEBUG
level.

Switches instrument from test_fulfillment_form to test_order_survey
when looking for swab and send encounter to make the check
consistent with what is pulled for the encounter_date.

This commit also condenses some of the event type logic.

The 'Testing Determination - Internal' instrument does not capture a
date, so calculate encounter_date based on the repeat instance number
and the date that the first repeat instance represents for the project.