Move common REDCap ETL code to redcap.py
This commit moves common code from redcap_det_uw_reopening.py to
redcap.py so that the code can be reused for other REDCap ETL projects.

The work was primarily cut-and-paste, but some elements were
parameterized to accommodate differences between REDCap projects and
future needs. For example, in redcap.py, the internal system
("https://seattleflu.org" in redcap_det_uw_reopening.py) is passed in
as a parameter.