etl redcap-det: revert REVISION number for REDCap DET ETLS
Due to concerns of very large reprocessing of data from all REDCap DETs,
revert the REVISION number for ETLs that do not necessarily need
to ingest Encounter.meta.source for all records. We can backfill
`encounter.details._provenance` by splitting the `encounter.identifier`
or just bumping the REVISION at a later date so we are not dealing with
so many changes at once.

The only ones we absolutely need to bump REVISION for are the
uw-retrospective and uw-reopening ETLs. uw-retrospective encounters have
a hashed `encounter.identifier` so it's not possible to backfill the
provenance data. We need to bump uw-reopening to ingest the skipped
`testing_determination_internal` encounters.