reportable_conditions notify: add notifications to new Childcare channel
hCoV-19 results for the Childcare also need to be sent to the new
`#ncov-reporting-childcare` channel because some members of the
Childcare study should not have access to the general `#ncov-reporting`
channel.

Due to time restrictions, this addition is done very lazily by
verifying all responses from Slack have a 200 status code. This does
mean that if one POST request fails, all of them will be sent again
on the next run of the notify command. The chance of this happening
seems pretty small, but we can revisit this at a later date if it
becomes a common issue.