Add valueDecimal support to a shipping view
In recent projects we started capturing questionnaire items as the
valueDecimal FHIR value type. Today, we're capturing `bmi` as valueDecimal.
We need to expose these in the shipping.fhir_questionnaire_responses_v1
view.

In python we capture these as `float`. In the shipping view we expose them
as the postgres `double precision` type. This type allows for 15 digits
of precision. We probably don't need that much precision for `bmi` but we
may for other items that get stored as valueDecimal.