etl redcap-det scan: Update zip code field handling
1. Update :func: `zipcode_map` to check if the provided *redcap_code* is
an emptry string to catch the edge case of when the zipcode field is
empty.

This field may be an empty string if edited by internal SFS members
for participants that live outside the SCAN coverage zip codes. We've
been doing in person dropoffs for priority codes that can be outside of
the normal SCAN coverage.

2. Replace the UnknownRedcapZipCode error with a warning that does not
stop the entire ETL pipeline. We want a notification when we run into
an unknown zip code, but it shouldn't stop the ingest process.

There's no need to update the geocoding process because SmartyStreets
can handle address lookups without a zipcode.