Add function for uw-reopening to insert synthetic DETs
This code is meant to be called from another simple script
in order to generate and insert syntethic DETs for UW Reopening
enrollment records. This is needed because there are so many DETs
that we do not want to bump the ETL revision number to process them all.
Instead, we will generate DETs for ranges of records via an external
script and "sprinkle" them into the database so that these DETs don't
delay real DETs from getting processed in a timely manner.