Log record info with uw-reopening ETL warnings
Currently, when we see a warning logged from the UW Reopening ETL, we have
to go to the syslog and grep through to find the warning entry and then
find the associated record number. This takes a lot of time. This commit
logs the record ID and in some places the record instance ID for record
related warnings.

In the future when we have a structured logging solution, we will have a
different way to get the record context with the event.