Add pre_analytical_specimen_collection column to results view
With the Adult Family Home Outbreak / Workplace Outbreak project,
we will now be collecting samples under a clinical context (in which
a clinical provider orders the test) instead of under an IRB context
(that does not require a provider's order). To the `shipping.return_results_v3`
view add a `pre_analytical_specimen_collection` column to distinguish this.