Childcare ETL: Use survey timestamp correctly
When using a REDCap server generated survey timestamp, we should do so
by using the `extract_date_from_survey_timestamp` function instead
of accessing the timestmap variable directly. If the survey has not
been completed, a value of `[not completed]` will be returned for
the timestamp. The function handles this correctly.