Add added_surveillance_groups into uw_priority_queue
For UW Reopening (HCT), a new variable `added_surveillance_groups` has
been added to the enrollment questionnaire. The values for this
variable are the days of the week (lower case). We should invite
a participant if for `added_surveillance_groups` they have the value
that matches today's day of the week, unless they're also a UW Housing
resident.