Pull `testing_date` from the testing_determination_internal form
For the UW-reopening ETL, by pulling the `testing_date` from the
testing_determination_internal form we will be able to give a more accurate
analysis of how much time passes between invitations for a given
participant.

The revision number is not incremented in this commit because all
records will be pulled manually.

I removed a sentence from the code comment that was not accurate.