Add view shipping.genome_submission_metadata_v1
This view will be used to export minimal metadata for submissions of
consensus genomes to external databases such as GISAID and GenBank.
We use the PUMA to link to the county, but if not available,
we just report the state, which is 'Washington' by default. The source
is used to link the consensus genome to the appropriate authors and
to report SARS-CoV-2 variants of concern to the appropriate study teams.
Swab type is nice to have, but is optional for submissions.