Test against Python 3.7, 3.8, and 3.9 in CI
Ensures that our code functions on these versions as we go along, so
that we can upgrade the deployed Python version in production easier in
the future.

Ditches Pipenv as that requires a single Python version (3.6).  It's
appropriate for a common team development environment, but not CI.

Adjusts ./dev/mypy to not run Pipenv when in CI.  It's still useful to
have a single way to run Mypy that DTRT.