Add shipping.scan_hcov19_result_counts_v2 view for SCAN dashboard
In order to exclude kits that have been shipped but not yet tested
from positivity counts on the SCAN dashboard, it is necessary to
add the priority code field to the view so we can differentiate
between those kits that have been confirmed positive and those not
yet tested.

An API endpoint will also need to be added to use this new view.