Tear down the hCoV-19 wall implemented by row-level security policies
In early March 2020, we built a wall in our production ID3C database
instance between "hCoV-19" (SARS-CoV-2, COVID-19) test results and
everything else.  This allowed us to grant only a small number of
internal study team members access to the SARS-CoV-2 test results.
These results were considered extremely sensitive at the time because of
their high public interest.

The wall persisted until now but is largely meaningless (and has been
for a while).  Much of the study team has visibility into SARS-CoV-2
test results and the results are no longer considered nearly as
sensitive.  The lab's testing and return of results to participants is
also on much firmer ground than it was back when the wall went up.

Tearing down the wall simplifies the internals of our production ID3C
database instance and will make data queries easier in Metabase.  It is a bit
of maintenance that's overdue.

These Sqitch reworks remove the functional partition of and restrictions
in the database.  An organizational partition still exists in Metabase,
which can be addressed separately at any future point.