backfill-sample-ids: New command to backfill sampleid for sample records
Add a new id3c command to mint new sample identifiers and assign them to
warehouse.sample records that are missing sample identifiers. This is useful
for tiny swabs samples, which do not have room on the tube for a sample
identifier, only a collection identifier. This process fills in the sample
record in our warehouse table but notably does not notify the lab of any
identifier assignments. We can use this backfilled sample ids as an
identifier when sharing data with third parties.