Fixup encounter and enrollment events in HCT shipping views
When we updated to include the new HCT server and project in
various shipping views, we mixed up encounter and enrollment
events for shipping.uw_reopening_encounters_v1 and
shipping.__uw_encounters. Fix these up so they are consistent
and we still pick up encounters from the old project as expected.