Skip HCT migrated records with no non-enrollment encounters
We are postponing processing of enrollment records for the HCT migrated
records until the data issues with those record are resolved in REDCap. This is
causing warnings when there are DETs being processed for migrated records'
enrollment forms that do not have any non-enrollment forms filled out.

Each FHIR bundle with a patient resource must contain at least one encounter resource,
which typically would be covered because the enrollment itself counts as an encounter.
Because those are not being processed for migrated records there needs to be at least
one non-enrollment encounter to create a valid FHIR bundle. This change checks to make
sure that is the case before generating the FHIR bundle.