Add UW Dropbox as collection method and site
For unobserved tiny swabs received via dropboxes, the ETL process is being updated
to use the Husky Test Kit Registration form to determine the collection method
and set the site value. This will allow the ETL process to link the specimen and
encounter entries so that ID3C will have link between sample and encounter.

There is no reliable way to determine if a self-registrered kit was received via
REDCap, since there is no PCDEQC form filled out for dropbox samples. This updated
process assumes that a specimen is received if the kit registration is complete,
which is not guaranteed but is likely the case for most samples.