Add check for multiple collection methods in repeat instance
This will skip processing of an encounter instance if it has more than one
collection method (kiosk, swab-n-send, dropbox) and triggers an alert so
that the data can be fixed in REDCap. Prior to this change, records that
had kiosk registration as well as one test order survey and/or husky test kit
registration form would only process the kiosk registration with no indication
it was skipping the others.