Add view permissions to return-results-exporter/grants
When deploying sqitch changes that reworked return-results-exporter/grants,
permissions on shipping views was dropped. These grants have been moved to
shipping/views sqitch file, so that when views are dropped and recreated the
permissions will be readded, but they are also needed here for cases
where grants file needs to be modified (i.e. for permissions on tables).