Fix encounter class mapper
The UW retrospectives ETL is pulling data from an EMR that doesn't
conform to the FHIR specification for encounter class, available at
https://www.hl7.org/fhir/v3/ActEncounterCode/vs.html.

In order to fix this, we set up a table that maps the EMR's values
to FHIR values, but it's missing several possible values. Add them
so the code stops throwing "Unknown encounter class" exceptions.