Check barcode before including specimen entry in FHIR bundle
Adding a check to determine if a barcode is valid before creating a
specimen entry in the FHIR bundle. This will prevent downstream warnings
in the FHIR ETL when an invalid barcode is encountered.

For now, this is only being implemented on UW reopening due to the higher
number of invalid barcodes received and the QC process transferring to LIMS.