HCT ETL: add daily_travel question, update study start date
Ingest in the new Y2 question daily_travel from redcap daily
attestation. This will be used in the quota system updates.

Also, update the study start date to the day we started redcap
project 45 on the new HCT redcap server. This date is used by
the ETL as a fallback when it can't find other dates, but we want
it to be updated to reflect the correct date to repeat instance
calculation.

As is, having one study start date in this code could lead to
inconsitency if we tried to process from both projects as once,
since they now have different start dates for the purposes of
calculating date from repeat instance.