Update regex on collapse columns function
Regex pattern on the collapse_columns function was incorrectly
defined allowing the last character of the `stub` parameter to be
optional. This had the effect of unintentionally collapsing `census_tract`
column when trying to collapse `race` columns because both contain `rac`.

The updated pattern requires all characters of the `stub` parameter to
be present in the column name.