Update SCH parsing and clincial ETL to include ICD-10 codes
Update parsing for SCH clinical data files to include ICD-10 codes. The
codes are stored in a series of 10-12 new columns in the workbooks for
years 1-3 and are being converted to an array to be included in the NDJSON
output.

The clinical ETL is updated to include this data during ingestion if it
is present.