Update SCH parsing to account for changes in Year 3 format
The flu shot column was renamed in year 3 data file from SCH, and a
flu shot date column was added. COVID vaccination (dose 1 and 2) columns
are also new, each with it's own date field, plus a single column for
COVID vaccine manufacturer.

These columns are being added for parsing to be included in the NDJSON
output.