Add new values to discharge disposition and test result mapping
Recently received data includes new values in the test result and discharge disposition
columns. These values are being mapped to standardized values for inclusion in the FHIR bundle.

ED dismiss is short for Emergency Department dismissal, and LWBS is Left Without Being Seen. Both
ED dismiss codes are being classified as other (`oth`) because they do not fit into any of the
FHIR discharge disposition values we are mapping to.