Redesign of shipping.sample_with_best_available_encounter_data_v1
This view suffers from poor performance due to the way it is using regex
to match samples to sites. To reduce the number of rows that require regex
matching, this updated version joins the data in three steps, first with
exact matches on site id, then on LIKE matches, then finally on regex matches,
with each iteration only joining to the records that haven't already found a
match.

Also adding Y4 to calculated `season` column since we are now in year 4.