Update to redesign of shipping.sample_with_best_available_encounter_data_v1
A recent rework of `shipping.sample_with_best_available_encounter_data_v1` (#274) incorrectly
applied `similar_escape` when using a LIKE comparison. This did not impact the results
of the query because it was followed up with a regex comparison that caught anything not
already matched. This change yields the same result but performs better since there
are fewer records that need to matched using regex.

Also lowercasing the `site_manifest_details` to effectively ignore case when comparing to
`manifest_regex` (which is already lowercase)