Handle unknown discharge disposition values
Commenting out this exception because until the Codebook is defined in REDCap to limit values
we will continue receiving unmapped values for discharge disposition. For now we will map
unknown values to `None`. After the Codebook is updated to limit values for this field
we will add them to the map and reimplement this exception.