Minor change to fhir ETL to unpin mypy
We started receiving an incompatible type error when we upgraded mypy previously,
so it was pinned to 0.910. The error was caused by mypy not being able to
determine the correct type for an optional nested dict that is part of a larger
dictionary being returned from a function. Initializing this key/value as an
empty dictionary prevents the incompatible type error in mypy.