Update gitignore to ignore data files
Ignore CSV, NDJSON, XLS, and XLSX files when pushing to this repo. There are
currently two such files in the `locations/data` directory, neither of which
has ever been updated since it was initially pushed. If these files do need to
be updated in the future they can be explicitly excluded in `.gitignore` or
forcefully added and pushed.