Add Cascadia Identifiers to Backfill Sample IDs
When the Cascadia identifier set was created, the identifier set
was not added to the backfill sample ids job. This change adds
that identifier set name so that null identifiers get backfilled
for this collection. I'll also go ahead and add some documentation
around this task so that it is clearer you need to do this when
new tiny collection sets are created.

This was tested by running `backfill-sample-ids` locally and making
sure that samples that were a part of the `collections-cascadia-tiny-swabs-home`
got pulled in to be backfilled. Since there are currently no Cascadia
samples, I just created one with a `null` identifier and tested that
it got pulled in to be backfilled, which it did.